Create a new example to test CMSIS fully-connected op
The new example is based on 'micro/kernels/fully_connected_test.cc' and
allows to test the CMSIS-NN implementation of fully-connected op.
Make sure to use the "cmsis-nn" tag (TAGS="cmsis-nn") when compiling
the project. For example in order to compile it as an Mbed project use
the following command:
$ make -f tensorflow/lite/micro/tools/make/Makefile TARGET=mbed TAGS="cmsis-nn" generate_fully_connected_cmsis_test_mbed_project